requirements.txt
pyserial>=3.5
matplotlib>=3.6.2
matplotlib-inline>=0.1.3
pandas>=1.3.4